refactor(sync): type reqConcurrent as *int64 to match regclient
ReqConcurrent exists to be assigned to regclient's config.Host.ReqConcurrent,
which is an int64, so declaring it as *int only buys a conversion at the
assignment. Matching the width drops the cast and removes the question of what
happens to a value that does not fit an int on a 32-bit build. The generated
config schema is unaffected: both widths emit "integer".

+++ pkg/extensions/sync/service.go
@@ -979,7 +979,7 @@ func newClient(opts syncconf.RegistryConfig, credentials syncconf.CredentialsFil
 	// mirror, so the effective cap is per host, not a shared total across the upstream and its mirrors.
 	// When unset, regclient's defaults are kept (ReqConcurrent=3, ReqPerSec=0 i.e. unlimited).
 	if opts.ReqConcurrent != nil {
-		hostConfig.ReqConcurrent = int64(*opts.ReqConcurrent)
+		hostConfig.ReqConcurrent = *opts.ReqConcurrent
 	}
 
 	if opts.ReqPerSec != nil {
